requirements.txt
appdirs==1.4.3
click==6.7
decorator==4.0.11
Django==1.11
django-annoying==0.10.3
django-bower==5.2.0
django-localflavor==1.4.1
django-registration==2.2
django-scheduler==0.8.3
django-secure==1.0.1
django-sslserver==0.19
geocoder==1.22.3
icalendar==3.11.3
olefile==0.44
packaging==16.8
Pillow==4.1.0
pkg-resources==0.0.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
ratelim==0.1.6
requests==2.13.0
simplejson==3.10.0
six==1.10.0
stripe==1.53.0
